source/requirements.txt
sgiref==3.5.2
Django==4.1.3
Pillow==9.3.0
psycopg2-binary
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.4
django-bootstrap-v5==1.0.11
django-phonenumber-field==7.0.1
django-phonenumbers
weasyprint==57.1
gunicorn
xlwt
djangorestframework==3.14.0
django-modeltranslation==0.18.9

flake8
flake8-bugbear
flake8-builtins
pep8-naming
flake8-commas
flake8-variables-names
flake8-import-order
isort
